better mechanism for comment persistence (#17)
It is now done with a help of Django caching mechanism. In
form_valid method of related view text of comment is md5-digested
and prepended to a list of currently "recorded" comments (it can have
at most 100 elements and has a TTL of 10 minutes). The list is stored
in Django cache, for production it's django-uwsgi-cache, for dev it's
memory local cache (we can't use this one for production because
workers don't share it). The list of hashes is transformed into JSON
and put on assignment page in a "script" tag (with a help of JS global
variable). Therefore, client-side JS can get the list and remove
entries with matching hashes from localStorage.